GH-79 code clean up

+++ core/core-awt/src/main/java/org/icepdf/core/io/RandomAccessFileInputStream.java
@@ -93,7 +93,7 @@ public class RandomAccessFileInputStream extends InputStream implements Seekable
 
     public long skip(long n) throws IOException {
         int nn = (int) (n);
-        return (long) m_RandomAccessFile.skipBytes(nn);
+        return m_RandomAccessFile.skipBytes(nn);
     }
