+++ core/core-awt/src/main/java/org/icepdf/core/io/SeekableInputConstrainedWrapper.java
@@ -79,7 +79,7 @@ public class SeekableInputConstrainedWrapper extends InputStream {
 
         if (remain <= 0)
             return -1;
-        length = (int) Math.min(Math.min(remain, (long) length), (long) Integer.MAX_VALUE);
+        length = (int) Math.min(Math.min(remain, length), Integer.MAX_VALUE);
         return streamDataInput.read(buffer, offset, length);
     }
 
@@ -102,7 +102,7 @@ public class SeekableInputConstrainedWrapper extends InputStream {
         long remain = getBytesRemaining();
         if (remain <= 0)
             return -1;
-        n = (int) Math.min(Math.min(remain, n), (long) Integer.MAX_VALUE);
+        n = (int) Math.min(Math.min(remain, n), Integer.MAX_VALUE);
         return streamDataInput.skip(n);
     }
